+++ docs/Changelog.md
@@ -1,7 +1,8 @@
 # Changelog
 
 # 0.2.20
-  - add `within` logic to timeframed pass-fail criterias 
+  - add `within` logic to timeframed pass-fail criterias
+  - added `use-dns-cache-mgr` option.
 
 # 0.2.19
   - fixed bug when in distributed tests VU count was not calculated properly.