+++ docs/Changelog.md
@@ -3,6 +3,7 @@
 # 0.2.20
   - add `within` logic to timeframed pass-fail criterias
   - added `use-dns-cache-mgr` option.
+  - default-domain option renamed to default-address (scheme, hostname, port are now parsed from this option).
 
 # 0.2.19
   - fixed bug when in distributed tests VU count was not calculated properly.