+++ docs/Changelog.md
@@ -1,7 +1,9 @@
 # Changelog
 
 # 0.2.20
-  - add `within` logic to timeframed pass-fail criterias 
+  - add `within` logic to timeframed pass-fail criterias
+  - added `use-dns-cache-mgr` option.
+  - default-domain option renamed to default-address (scheme, hostname, port are now parsed from this option).
 
 # 0.2.19
   - fixed bug when in distributed tests VU count was not calculated properly.